requirements.txt
asgiref==3.7.2
Django==5.0
django-enumfields==2.1.1
psycopg2-binary==2.9.9
sqlparse==0.4.4
